Move IO into utils.
